renamed package to com.rocketchat.sdkdemo
