+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion "23.0.3"
     defaultConfig {
-        applicationId "com.example.rocketchat"
+        applicationId "com.rocketchat.sdkdemo"
         minSdkVersion 15
         targetSdkVersion 25
         versionCode 1
