+++ app/src/main/AndroidManifest.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.rocketchat">
+    package="com.rocketchat.sdkdemo">
 
     <application
-        android:name=".MyApplication"
+        android:name="com.rocketchat.sdkdemo.MyApplication"
         android:allowBackup="true"
         android:icon="@mipmap/rocketchat_logo"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.rocketchat.sdkdemo.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
