+++ app/src/main/java/com/rocketchat/sdkdemo/MyApplication.java
@@ -1,4 +1,4 @@
-package com.example.rocketchat;
+package com.rocketchat.sdkdemo;
 
 import com.github.rocketchat.livechat.Application.LiveChatApplication;
